% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argiope.R
\docType{package}
\name{Argiope}
\alias{Argiope}
\title{Image colour standardisation and data extraction.}
\description{
Converts image to standardised RGB values based on a known
grey scale. Use AI to standardise images and extract colorimetric data or interactively extract data from standardised images.
}
\author{
Felipe M. Gawryszewski \email{f.gawry@gmail.com}
}
