% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourchecker predictions.R
\name{colorchecker.scale.data.extraction}
\alias{colorchecker.scale.data.extraction}
\title{Extract grey values using AI}
\usage{
colorchecker.scale.data.extraction(
  image.model,
  image.raw,
  model,
  save.colorchecker.image = TRUE,
  save.name,
  colorchecker_y_lim,
  colorchecker_x_lim,
  colorchecker_y_size,
  colorchecker_x_size
)
}
\arguments{
\item{image.model}{Array of image to apply the AI model, same size as model training. Usually a 512x512px or 256x256px image.}

\item{image.raw}{The raw version of the 'image.model'. Usually a larger and with different proportions. Those are adjusted internally to proper data extraction.}

\item{model}{The Keras Unet model used for predictions.}

\item{save.colorchecker.image}{Whether the extracted colorchecker image should be saved.}

\item{save.name}{Name of file to be saved.}

\item{colorchecker_y_lim}{Relative position of the bottom and top vertex of each grey scale value to be extracted. Used when method = "AI".}

\item{colorchecker_x_lim}{Relative position of the left and rigth vertex of each grey scale value to be extracted. Used when method = "AI".}

\item{colorchecker_y_size}{The size of the colorchecker rectangle minor side.}

\item{colorchecker_x_size}{The size of the colorchecker rectangle largest side.}
}
\value{
A list containing: 1. Median of R,G, and B scale values; 2. MAD of R,G, and B pixel scale values; and 3. Image scale in pixel/cm2 and pixel/cm. Those are calculated based on the colorchecker size.
}
\description{
Extract grey values from the colourchecker using an AI model. This function is used internally.
}
