% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argiope_functions.R
\name{dcraw.tags}
\alias{dcraw.tags}
\title{Extract dcraw metadata}
\usage{
dcraw.tags(files)
}
\arguments{
\item{files}{A vector of image file names.}
}
\value{
A data.frame with metadata.
}
\description{
Get camera metadata using dcraw.
}
\details{
This function uses dcraw by Dave Coffin (https://www.dechifro.org/dcraw/). Windows version was compiled by Sergio Namias (http://www.centrostudiprogressofotografico.it/en/dcraw/). This is a 64bit version. It may not work in Windows 32bit OSs.
}
