% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{find.multimax}
\alias{find.multimax}
\title{Find the most common RGB pixel triplet values.}
\usage{
find.multimax(matrix)
}
\arguments{
\item{matrix}{An array with R, G and B values.}
}
\description{
Find the most common RGB pixel triplet values using a Kernel density estimation.
}
