% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argiope_functions.R
\name{find_group}
\alias{find_group}
\title{Find groups}
\usage{
find_group(file.name, interval = 3)
}
\arguments{
\item{file.name}{A vector of image file names.}

\item{interval}{Interval, in seconds, to group images.}
}
\value{
A data.frame with selected exif metadata and groupings.
}
\description{
Find group of photos based on exif data and photo interval time, in seconds. This function is used internally.
}
