% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argiope_functions.R
\name{get_scale_interactive}
\alias{get_scale_interactive}
\title{Find grey values coordinates interactively}
\usage{
get_scale_interactive(img, crop, N)
}
\arguments{
\item{img}{A tiff image generated by dcraw.}

\item{crop}{crop percentage to reduce image size at each side. e.g. crop = c(10,10,10,10).}

\item{N}{Number of points.}
}
\description{
Find grey values coordinates interactively. This function is used internally.
}
