% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argiope_functions.R
\name{get_scale_values}
\alias{get_scale_values}
\title{Extract grey values}
\usage{
get_scale_values(
  img,
  method = c("AI", "interactive"),
  model,
  image.model,
  colorchecker_x_lim,
  colorchecker_y_lim,
  colorchecker_y_size,
  colorchecker_x_size,
  X,
  Y,
  Size,
  save = TRUE,
  filename
)
}
\arguments{
\item{img}{The raw version of the 'image.model'. Usually a larger and with different proportions. Those are adjusted internally to proper data extraction. A tiff image generated by dcraw.}

\item{method}{Whether a UNET has been used to find the colorcheker scale or select scale location interactively.}

\item{model}{The Keras Unet model used for predictions. Used when method = "AI".}

\item{image.model}{Array of image to apply the AI model, same size as model training. Usually a 512x512px or 256x256px image.}

\item{colorchecker_x_lim}{Relative position of the left and rigth vertex of each grey scale value to be extracted. Used when method = "AI".}

\item{colorchecker_y_lim}{Relative position of the bottom and top vertex of each grey scale value to be extracted. Used when method = "AI".}

\item{colorchecker_y_size}{The size of the colorchecker rectangle minor side.}

\item{colorchecker_x_size}{The size of the colorchecker rectangle largest side.}

\item{X}{The x-axis center point of colorcheker scale squares. Used when method = "interactive".}

\item{Y}{The y-axis center point of colorcheker scale squares. Used when method = "interactive".}

\item{Size}{The size, in pixels, of the square to get colorcheker scale values. Used when method = "interactive".}

\item{save}{Should the colorchecker image be saved?}

\item{filename}{Name of file to be saved.}
}
\value{
A list containing: 1. Median of R,G, and B scale values; 2. MAD of R,G, and B pixel scale values; and 3. Image scale in pixel/cm2 and pixel/cm. Those are calculated based on the colorchecker size (available when 'method = "AI"' only).
}
\description{
Extract grey values from the colourchecker. This function is used internally.
}
