% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argiope_functions.R
\name{getPhotoData}
\alias{getPhotoData}
\title{Interactively get photo data}
\usage{
getPhotoData(
  method = c("file", "pred"),
  scale = NULL,
  path = getwd(),
  path.spec = getwd(),
  files = list.files(path = path, pattern = ".tif", ignore.case = TRUE),
  associate.spec = TRUE,
  spec.files = list.files(path = path.spec, pattern = ".txt", ignore.case = TRUE),
  order = TRUE,
  factor = 10,
  crop = 0,
  item.location = c("Mata A", "Mata B", "Mata C", "Cerrado A", "Cerrado B", "Cerrado C"),
  item.types = c("folha cima", "folha baixo", "folha outro", "casca", "casca liquen",
    "graminea verde", "graminea seca", "serrapilheira", "animal dorso", "animal ventre",
    "animal lateral")
)
}
\arguments{
\item{method}{Get data from a converted tif file ('file') or process RAW photo on the fly ('pred')}

\item{scale}{Object from the procphotos functions. Used when method='pred'.}

\item{path}{Path to access converted photos}

\item{path.spec}{Path to access reflectance files}

\item{files}{Vector of file names to analyze. When method='pred', use names(scale$scale) to get only the files selected by the procphotos functions.}

\item{associate.spec}{Logical. Should photos be associated with spec files.}

\item{spec.files}{Vector of spec names to associate with photos}

\item{order}{Order files by name?}

\item{factor}{factor to reduce image size}

\item{crop}{crop percentage to reduce image size}

\item{item.location}{A vector of item locations to be selected by user.}

\item{item.types}{A vector of item types to be selected by user.}
}
\value{
A data frame in each lines represent data of one extracted patch.
 One image of each patch is also saved. Results and images are saved in a 'results' folder.
}
\description{
Interactively get data from a converted photo and associate it with item type, item name and reflectance file.
}
