% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourchecker predictions.R
\name{getPhotoDataAI}
\alias{getPhotoDataAI}
\title{Get photo data using AI}
\usage{
getPhotoDataAI(
  method = c("file", "pred"),
  scale = NULL,
  scale.size = scale,
  path,
  path.image.model,
  model,
  item.name = "butterfly",
  files = "All"
)
}
\arguments{
\item{method}{Get data from a converted tif file ('file') or process RAW photo on the fly ('pred')}

\item{scale}{Object from the procphotos functions. Used when method='pred'.}

\item{scale.size}{Object from the procphotos functions used to set the scale of the image (i.e. pixel/cm2). It must be provided.}

\item{path}{Path to access raw photos.}

\item{path.image.model}{Path to access images to apply the AI model, same size as model training. Usually a 512x512px or 256x256px image.}

\item{model}{The Keras Unet model used for predictions.}

\item{item.name}{A vector of item types to be selected by user.}

\item{files}{Vector of file names to analyze. If 'files="All"', all files are processed.}
}
\value{
A data frame in each lines represent data of one extracted patch.
 One image of each patch is also saved. Results and images are saved in a 'results' folder.
}
\description{
Get data from a converted photo using an AI model.
}
