% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk.R
\name{item.name.fun}
\alias{item.name.fun}
\title{Ask user}
\usage{
item.name.fun()
}
\value{
Prompt user and return value.
}
\description{
Function to ask item name to user. This function is used internally. Based on https://stackoverflow.com/questions/16847621/get-data-out-of-a-tcltk-function
}
