% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argiope_functions.R
\name{plot.procphotos}
\alias{plot.procphotos}
\title{Verify conversion.}
\usage{
plot.procphotos(procphotos, scales = "colorchecker")
}
\arguments{
\item{procphotos}{Results from procphotos function}

\item{scales}{Path to scales JPG images. Usually 'colorchecker' directory.}
}
\value{
A plot with scale image and regression data of each photo.
}
\description{
Plot to verify conversion from procphotos function.
}
