% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{plotIMG}
\alias{plotIMG}
\title{Plot image from R,G,B tif file}
\usage{
plotIMG(file, fix.values = TRUE)
}
\arguments{
\item{file}{A tiff image. Usually generated by the Argiope package.}

\item{fix.values}{Fix negative and >100% values for better presentation.}
}
\description{
Mostly for visualisation.
}
