% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argiope_functions.R
\name{predict_reflect}
\alias{predict_reflect}
\title{Extract grey values}
\usage{
predict_reflect(
  scale,
  img,
  whichRvalues,
  RvaluesR,
  RvaluesG,
  RvaluesB,
  return.image,
  save,
  file.name = NULL
)
}
\arguments{
\item{scale}{Scale data from function get_scale_values}

\item{img}{A tiff image generated by dcraw.}

\item{whichRvalues}{Which grey values should be used.}

\item{RvaluesR}{Reflectance values for the R channel.}

\item{RvaluesG}{Reflectance values for the G channel.}

\item{RvaluesB}{Reflectance values for the B channel.}

\item{return.image}{Logical. Should the image be returned?}

\item{save}{Logical. Should the converted image be saved?}

\item{file.name}{Image file name.}
}
\value{
A list with: 1. linear models used to convert images; and 2. the converted image.
}
\description{
Calculates new R,G and B reflectance values for the image based on scale values
}
