% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argiope_functions.R
\name{procphotos}
\alias{procphotos}
\title{Convert photos}
\usage{
procphotos(
  auto = TRUE,
  interval = 3,
  limit = 0.8,
  path,
  ext = ".CR2",
  method = c("AI", "interactive"),
  model,
  path.image.model,
  ext.model = ".JPG",
  colorchecker_y_lim = c(0.25, 0.75),
  colorchecker_x_lim = c(0.04, 0.11, 0.21, 0.28, 0.38, 0.45, 0.55, 0.62, 0.72, 0.79,
    0.89, 0.96),
  colorchecker_y_size = 1.223,
  colorchecker_x_size = 8.738,
  crop = c(0, 0, 0, 0),
  Size = NULL,
  RvaluesR = c(98.35, 65.22, 41.42, 23.17, 9.97, 2.17),
  RvaluesG = c(97.89, 66.19, 42.45, 23.69, 10.48, 2.22),
  RvaluesB = c(95.98, 66.87, 43.3, 24.42, 11.13, 2.26),
  whichRvalues = c(1:6),
  bug.rotate = FALSE,
  save = FALSE
)
}
\arguments{
\item{auto}{Logical. Should best photos be selected automatically?}

\item{interval}{Interval, in seconds, to group images. Used when auto=TRUE.}

\item{limit}{A threshold of maximum grey scale reflectance value by which an image is rejected. Used when auto=TRUE.}

\item{path}{Path containing RAW images to be analysed.}

\item{ext}{Raw image extension.}

\item{method}{Whether a UNET AI has been used to find the colorcheker scale or select colorchecker interactively.}

\item{model}{The Keras Unet model used for predictions. Used when 'method = AI'.}

\item{path.image.model}{#Path containing images to be predicted by model, same size as model training. A smaller version of 'img'. Used when 'method = AI'.}

\item{ext.model}{Unet model image extension. Used when 'method = AI'.}

\item{colorchecker_y_lim}{Relative position of the bottom and top vertex of each grey scale value to be extracted. Used when 'method = AI'.}

\item{colorchecker_x_lim}{Relative position of the left and right vertex of each grey scale value to be extracted. Used when 'method = AI'.}

\item{colorchecker_y_size}{The size of the colorchecker rectangle minor side.}

\item{colorchecker_x_size}{The size of the colorchecker rectangle largest side.}

\item{crop}{Crop percentage to reduce image size in each side. e.g. crop = c(10,10,10,10).}

\item{Size}{The size, in pixels, of the square to get colorcheker scale values. Used when 'method = interactive'.}

\item{RvaluesR}{Reflectance values for the R channel.}

\item{RvaluesG}{Reflectance values for the G channel.}

\item{RvaluesB}{Reflectance values for the B channel.}

\item{whichRvalues}{Which grey values should be used.}

\item{bug.rotate}{Should image be rotate in case of low R2?}

\item{save}{Logical. Should the converted image be saved? Very large TIFF files.}
}
\value{
A list containing: 1. Median of R,G, and B scale values; 2. R,G, and B pixel scale values; and 3. the linear models used to convert images;
}
\description{
Convert photos to standardised R, G and B values based on a known grey scale.
}
\details{
This function uses dcraw by Dave Coffin (https://www.dechifro.org/dcraw/). Windows version was compiled by Sergio Namias (http://www.centrostudiprogressofotografico.it/en/dcraw/). This is a 64bit version. It may not work in Windows 32bit OSs. On Macs 'dcraw' should be installed manually: https://formulae.brew.sh/formula/dcraw
}
