% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argiope_functions.R
\name{read.dcraw}
\alias{read.dcraw}
\title{Read RAW photos}
\usage{
read.dcraw(file, args = "-W -o 0 -q 0 -4 -T")
}
\arguments{
\item{file}{Raw file name.}

\item{args}{arguments to be passed to dcraw.}
}
\value{
An array with RGB values.
}
\description{
Get camera data using dcraw.
}
\details{
This function uses dcraw by Dave Coffin (https://www.dechifro.org/dcraw/). Windows version was compiled by Sergio Namias (http://www.centrostudiprogressofotografico.it/en/dcraw/). This is a 64bit version. It may not work in Windows 32bit OSs. Code was based on read.raw function in package Adimpro. To install dcraw on Macs see https://r-spatial.org/r/2023/05/15/evolution4.html. Currently it works only with tiff generated files (i.e. args must include -T dcraw parameter).
}
