% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_jpeg_for_AI.R
\name{resize_for_AI}
\alias{resize_for_AI}
\title{Padding and resize images for AI}
\usage{
resize_for_AI(geometry = "512x512", pattern = ".jpg", path.images)
}
\arguments{
\item{geometry}{Image conversion size.}

\item{pattern}{Image extension.}

\item{path.images}{Path to access photos.}
}
\description{
Interactively get data from a converted photo and associate it with item type, item name and reflectance file.
}
