% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{SDimages}
\alias{SDimages}
\title{Extract colour data from tif images.}
\usage{
SDimages(files)
}
\arguments{
\item{files}{A vector with tif image names.}
}
\description{
Extract and calculates several colour metrics from tif images.
}
