% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argiope_functions.R
\name{Wiener.training}
\alias{Wiener.training}
\title{Wiener training}
\usage{
Wiener.training(R = NULL, tRGB = NULL)
}
\arguments{
\item{R}{A data.frame of training reflactance curves. Do not include a column with wavelength values.}

\item{tRGB}{A matrix with R, G, and B values (columns) associated with the reflectance values}
}
\value{
The G matrix
}
\description{
Find the G matrix based on a training RGB and reflectance data.
}
